Update imf.py
Previous behaviour would basically recreate the load indicator functionality, which made it pointless to have loaded the indicator, since the file would be read from disk again (and there was no failsafe if a user called to get data before loading it).

This matches the behaviour elsewhere, which is to unpack the dictionary into a dataframe in long format.

@lpicci96 for info, please have a look at the new implementation